requirements.txt
asgiref==3.6.0
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-cors-headers==4.0.0
django-countries-plus==2.1.0
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
pytz==2023.3
requests==2.30.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
webencodings==0.5.1
whitenoise==6.4.0
boto3==1.26.152
botocore==1.29.165

